Add priority to the 'widget_display_callback' filter
Set a priority for the widget_display_callback filter so that other plugins can also short circuit widgets by coming afterwards.  Because show_widget returns an $instance that is not null if it does not match, if it comes after the other plugin's filter, the widget will still display, even if the first filter returns false.